Add files via upload
Write a smart contract using Ethereum solidity for taking attendance for a single class on a particular day. The contract should have the following features:
The address that creates the contract will be the address of the course instructor
When creating the contract, the instructor can pass the course ID and the date for which he/she wants to take attendance as two-string arguments. That is, the constructor function in the contract should take these two arguments and store them in the contract.
After creating the contract, the instructor can invoke a function to enable and then disable attendance taking. These two functions can be invoked by the instructor only. If any other address tries to invoke these functions, the contract should throw an error.
Any address (representing a student) can invoke a function give-attendance when attendance taking is enabled. In that function, the student can pass his/her student ID. The function will record both the address and the student ID in an attendee list when this function is invoked. The same address cannot invoke give-attendance more than once.
Any address can check if it has given attendance using a check-attendance function. This function can be invoked any number of times. The function returns the student ID if the address gave-attendance before. Otherwise, it should return null.
There will be an alternative check-attendance function for the instructor. In this function, the instructor will pass the student-id and the function will return true or false depending on whether or not the student-id is on the attendee list. Note that the function should give an error if any address other than the instructor tries to invoke this function.
Finally, there will be a function to count the total number of attendance. Anyone can invoke this function.